// MyWXFrame/ui/src/hasssimple/common/types.ts
  export interface Credential {
    auth_provider_type: string;
    auth_provider_id: string;
  }
  
  export interface MFAModule {
    id: string;
    name: string;
    enabled: boolean;
  }
  
  export interface CurrentUser {
    id: string;
    is_owner: boolean;
    is_admin: boolean;
    name: string;
    credentials: Credential[];
    mfa_modules: MFAModule[];
  }
  
  // Currently selected theme and its settings. These are the values stored in local storage.
  export interface ThemeSettings {
    theme: string;
    // Radio box selection for theme picker. Do not use in cards as
    // it can be undefined == auto.
    // Property hass.themes.darkMode carries effective current mode.
    dark?: boolean;
    primaryColor?: string;
    accentColor?: string;
  }
  
  export interface PanelInfo<T = Record<string, any> | null> {
    component_name: string;
    config: T;
    icon: string | null;
    title: string | null;
    url_path: string;
  }
  
  export interface Panels {
    [name: string]: PanelInfo;
  }
// ...
  export interface Translation {
    nativeName: string;
    isRTL: boolean;
    hash: string;
  }
  
  export interface TranslationMetadata {
    fragments: string[];
    translations: {
      [lang: string]: Translation;
    };
  }
  
  export interface IconMetaFile {
    version: string;
    parts: IconMeta[];
  }
  
  export interface IconMeta {
    start: string;
    file: string;
  }
  
  export interface Notification {
    notification_id: string;
    message: string;
    title: string;
    status: "read" | "unread";
    created_at: string;
  }
  
  export interface Resources {
    [language: string]: Record<string, string>;
  }
  
  export interface Context {
    id: string;
    parent_id?: string;
    user_id?: string | null;
  }
  
  export interface ServiceCallResponse {
    context: Context;
  }
  
  export interface ServiceCallRequest {
    domain: string;
    service: string;
    serviceData?: Record<string, any>;
    target?: any;
  }
  
  export interface HomeAssistant {
    auth: Auth;// & { external?: ExternalMessaging };
    connection: Connection;
    connected: boolean;
    states: HassEntities;
    services: HassServices;
    config: HassConfig;
    themes: any;// Themes;
    selectedTheme: ThemeSettings | null;
    panels: Panels;
    panelUrl: string;
    // i18n
    // current effective language in that order:
    //   - backend saved user selected language
    //   - language in local app storage
    //   - browser language
    //   - english (en)
    language: string;
    // local stored language, keep that name for backward compatibility
    selectedLanguage: string | null;
    locale:any;//  FrontendLocaleData;
    resources: Resources;
    localize: any;// LocalizeFunc;
    translationMetadata: TranslationMetadata;
    suspendWhenHidden: boolean;
    enableShortcuts: boolean;
    vibrate: boolean;
    dockedSidebar: "docked" | "always_hidden" | "auto";
    defaultPanel: string;
    moreInfoEntityId: string | null;
    user?: CurrentUser;
    userData?: any | null;
    hassUrl(path?:any): string;
    callService(
      domain: ServiceCallRequest["domain"],
      service: ServiceCallRequest["service"],
      serviceData?: ServiceCallRequest["serviceData"],
      target?: ServiceCallRequest["target"]
    ): Promise<ServiceCallResponse>;
    callApi<T>(
      method: "GET" | "POST" | "PUT" | "DELETE",
      path: string,
      parameters?: Record<string, any>,
      headers?: Record<string, string>
    ): Promise<T>;
    fetchWithAuth(path: string, init?: Record<string, any>): Promise<Response>;
    sendWS(msg: MessageBase): void;
    callWS<T>(msg: MessageBase): Promise<T>;
    // loadBackendTranslation(
    //   category: Parameters<typeof getHassTranslations>[2],
    //   integration?: Parameters<typeof getHassTranslations>[3],
    //   configFlow?: Parameters<typeof getHassTranslations>[4]
    // ): Promise<LocalizeFunc>;
    // loadFragmentTranslation(fragment: string): Promise<LocalizeFunc | undefined>;
  }